Rename to rocto-client, update README
