+++ rocto_client/Qt/ui/auto_pyuic.sh
