+++ rocto_client/Qt/ui/source/preferencesdialog.ui
