+++ rocto_client/Qt/ui/source/settingsdialog.ui
